backend/requirements.txt
fastapi>=0.100.0
uvicorn==0.22.0
redis==4.5.4
elasticsearch==8.7.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
python-multipart==0.0.6
user-agents==2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
joblib
requests>=1.3.0
pyyaml>=6.0
